Common interfaces of DE

Core DE

The core data structure of differential evolution is a struct that inherits the abstract type Evolver or, more typically, the DEEvolver, which itself inherits Evolver and represents a single-population DE.

DifferentialEvolution.Evolver — Type

Evolver

Abstract type of all DE core.

DifferentialEvolution.Options — Type

Options

Abstract type of all DE options.

DifferentialEvolution.DEEvolver — Type

DEEvolver <: Evolver

An abstract normal (single-population) DE evolution type

DifferentialEvolution.DEOptions — Type

DEOptions <: Options

Abstract type of options for a DE, particually, a single-population one. See also DEEvolver.

Each concrete type of DEEvolver should support the five default interfaces listed below. If the five corresponding fields are defined, then nothing extra needs to be done.

pop(e::DEEvolver) = e.pop
fitness(e::DEEvolver) = e.fitness
senses(e::DEEvolver) = e.senses
bounds(e::DEEvolver) = e.bounds
options(e::DEEvolver) = e.options

Thus, the five fundamental fields listed above can be retrieved directly or using a method with the same name of the field.

DifferentialEvolution.pop — Function

pop(e::DEEvolver)

Get the population of a DE, which is typically a matrix.

DifferentialEvolution.fitness — Function

fitness(e::DEEvolver)

Get the fitness of the population, which is generally a vector for single-objective DE or a matrix for multi-objective DE.

DifferentialEvolution.senses — Function

senses(e::DEEvolver)

Get the senses of optimization. In traditional evolutionary algorithms, the fitness of an individual is going to be maximized. If minimization is required instead, then set a negative sense.

DifferentialEvolution.bounds — Function

bounds(e::DEEvolver)

Get the bound constraints,typically a $n x 2$ matrix for n variables.

DifferentialEvolution.options — Function

options(e::DEEvolver)

Get the options of this DE.

DE options

Options, i.e., control parameters, of DE are encapsulated in a struct as a subtype of DEOptions (for the most common single-population DE) or a subtype of Options. The options of a given DE is accessed by the method options.

Other utility interfaces

The following three methods have default implementation for DEEvolver.

DifferentialEvolution.nindividuals — Method

nindividuals(e::DEEvolver)

Number of individuals in the population.

Base.ndims — Method

ndims(de::DEEvolver)

Number of variables (dimension) of the DE.

DifferentialEvolution.nobjectives — Method

nobjectives(e::Evolver)

Number of objectives.

No default implementation is provided for the two methods below, since a correct implementation depends on details of the concrete DE type.

DifferentialEvolution.best_individual — Method

best_individual(e::Evolver; all=false)

Get individuals that have the best fitness. If all is true, then all such individuals are returned; otherwise, only one of them is fetched.

DifferentialEvolution.best_fitness — Method

best_fitness(e::Evolver)

Get the best fitness among the individuals of the evolver.

It is highly recommended that a concrete DE type implements the two best methods as outlined above.

Currently, DifferentialEvolution.jl provides a standard implementation with the types StdDE and StdOptions.