Crossover

Crossover (also known as mating or recombination) operations in DE. Note that we recommend in-place crossover operators: supposing two vectors X and V are recombined into another vector U, the vector U should be provided as the first argument and be filled inside the method.

Binomial crossover

DifferentialEvolution.crossover_binomial! — Function

crossover_binomial!(U::AV{T}, i::Integer, de::StdDE{<:AM{T}}, V::AV{T}) where T<:AF

Recombine the i-th vector of de and the donor vector V into a trial vector U. The crossover rate is specied by de.options.

See also StdDE.

crossover_binomial!(U::AV{T}, X::AV{T}, V::AV{T}, Cr::AF) where T<:AF

Recombine two vectors X and V into U subject to the rate Cr.