Mutation

Common mutation operators in DE are provided. We recommend in-place operators for efficiency purpose. That is, a dummy donor vector is the first argument, which is then filled inside the method.

DE/rand/1

DifferentialEvolution.mutate_rand_1! — Function

mutate_rand_1!(V::AV{T}, i::Integer, de::StdDE{<:AM{T}}) where T<:AF

DE/rand/1 mutation in place, which produces a donor vector V to crossover with the i-th target vector (parent).

mutate_rand_1!(V::AV{T}, i::Integer, pop::AM{T}, F::AF, indices::AV{<:Integer}) where T<:AF

A lower-level implementation. V is the donor vector to be filled, i is the index of the target vector in the population pop whose size is $n$, F is the scale factor, and indices is an array of length $n$ that contains intergers from 1 to $n$ (in an arbitrary order).