Standard DE

The standard single-objective DE types and evolution algortithms.

Types

DifferentialEvolution.StdOptions — Type

StdOptions{T<:AF}

Standard options with two parameters, F (scaling) and Cr (crossover rate). This type can be constructed with either positional arguments or keyword arguments.

DifferentialEvolution.StdDE — Type

StdDE{TP<:AM{<:AF}, TF<:AVoM{<:Real}, TS<:ASoV{<:Integer}, TO<:DEOptions} <: DEEvolver

The standard DE type with the following fields:

pop::TP: a matrix-type population, each column of which is an individual (vector) in DE

fitness::TF: fitness of each individual. For single-objective DE, fitness is a vector-like structure; for multi-objective DE, fitness is of a matrix type, whose column contains values of multiple objectives for each individual.

senses::TS: integer for a single objective or vector of integers for multiple objectives. An individual's strength is defined by fitness times senses. Only values 1 (maximization) and -1 (minimization) are allowed. To perform minimization of an objective, set a negative sense, i.e., -1.

bounds::TP: a d-by-2 matrix for a d-dim DE, each row containing the lower and upper boundaries of a variable.

options::TO: options which usually contain the control parameters. See also StdOptions.

In practice, more convenient types are usually used like StdSoDE for single-objective DE and StdMoDE for multiobjective DE.

DifferentialEvolution.StdSoDE — Type

StdSoDE{TP<:AM{<:AF}, TF<:AV{<:Real}, TW<:Real, TO<:DEOptions}

Standard single-objective DE.

DifferentialEvolution.StdMoDE — Type

StdMoDE{TP<:AM{<:AF}, TF<:AM{<:Real}, TW<:AV{<:Real}, TO<:DEOptions}

Standard multi-objective DE.

Constructor

The three standard DE types StdDE, StdSoDE, and StdMoDE can all be built with their inner constructors, but a more convenient way is to employ the dedicated outer constructor. Please check Julia documentation on Constructors for more details.

DifferentialEvolution.StdDE — Method

StdDE(::Type{TF}, nindividuals::Int, bounds::AM{<:AF}, opt::DEOptions; weights::ASoV{<:Real}) where {TF<:Real}

Create a standard DE evolver with a population of nindividuals individuals and options specified by opt. The length of each individual (also known as a vector) is equal to the number of rows of bounds. That is, each row of bounds denotes the range $[l, u]$ of a variable, where $l$ and $u$ denote the lower and upper bound, respectively.

TF is the data type of fitness. The type of individuals (solution vectors) is the same as the type of bounds.

The size of senses indicates the number of objectives. If the supplied senses is a scalar, then a single-objective DE StdSoDE instance is built; otherwise, a multi-objective StdMoDE is constructed.

DifferentialEvolution.StdDE — Method

StdDE(nindividuals::Int, bounds::AM{<:AbstractFloat}, opt::DEOptions; senses::ASoV{<:Integer})

Construct a standard DE whose vectors are of default type Float64. See StdDE(::Type{TF}, nindividuals::Int, bounds::AM{<:AbstractFloat}, opt::DEOptions; senses::ASoV{<:Integer}) where {TF<:Real} for details.

Evolution

The basic single-objective DE evolution algorithm is implemented.

DifferentialEvolution.evolve! — Function

evolve!(de::StdDES, evaluator::Function, ngen; mut=mutate_rand_1, cx=crossover_binomial,
    bc=:bounce_back)

Standard single-objective differential evolution.

Each individual's fitness is evaluated by the given evaluator and the evolution lasts for ngen iterations. Keyword arguments:

mut=mutate_rand_1!: mutation method

cx=crossover_binomial!: crossover method, currently support

bc=:bounce_back: boundary constraints handling method， currently support

:bounce_back

stats=[:best, :avg]: fitness statistics to display in each generation. Valid entries are :best, :avg(average), :max, :min, :std(standard deviation).

stats_fmt::FormatSpec=FormatSpec("10.5e"): format specification of each statistic term for display. See Formatting.jl for details.

Ranking

DifferentialEvolution.best_individual — Method

best_individual(de::StdSoDE; all=false)

Get the best individual (or individuals) from the single-objective DE population. See also best_individual(e::Evolver; all=false) for more details.

DifferentialEvolution.best_fitness — Method

best_fitness(de::StdSoDE)

Get the best fitness from the single-objective DE population.

Base.sort! — Method

sort!(de::StdDES; rev::Bool=false)

Sort a single-objective DE population in place according to the fitness. The best individual appears at first if rev is false (the default value).